purchase screen update2

+++ OOP - Initial Prototype/Game.cpp
@@ -1,6 +1,6 @@
 #include "Game.h"
 
-Game::Game(int id, const std::string& name, const std::string& desc, int cost, int rating)
+Game::Game(int id, const std::string& name, const std::string& desc, float cost, int rating)
 	: id(id),name(name), description(desc), cost(cost), ageRating(rating)
 {
 }
@@ -19,7 +19,7 @@ const std::string& Game::GetDescription() const
 	return description;
 }
 
-int Game::GetCost() const
+float Game::GetCost() const
 {
 	return cost;
 }
