+++ OOP - Initial Prototype/Game.h
@@ -5,12 +5,12 @@
 class Game
 {
 	public:
-		Game(int ,const std::string&, const std::string&, int, int);
+		Game(int ,const std::string&, const std::string&, float, int);
 		~Game();
 
 		const std::string& GetName() const;
 		const std::string& GetDescription() const;
-		int GetCost() const;
+		float GetCost() const;
 		int getId();
 		int getAgeRating();
 
@@ -18,7 +18,7 @@ class Game
 		int id;
 		std::string name;
 		std::string description;
-		int cost;
+		float cost;
 		int ageRating;
 
 };
